% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{as.character.yac_symbol}
\alias{as.character.yac_symbol}
\title{Convert yac symbol to character}
\usage{
\method{as.character}{yac_symbol}(x, ...)
}
\arguments{
\item{x}{A \code{yac_symbol}}

\item{\dots}{not used}
}
\description{
Convert yac symbol to character
}
\concept{yac_symbol}
