% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sym2.R
\name{as.Sym.character}
\alias{as.Sym.character}
\title{Convert character vector to yacas object}
\usage{
\method{as.Sym}{character}(x, ...)
}
\description{
Simple and raw conversion to yacas
}
\examples{
x <- c("a", "2", "4", "c", "d", "6")
x
y <- as.Sym(x)
y
}
