% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_r}
\alias{as_r}
\title{Convert yacas vector (list) or matrix (list of lists) to R vector/matrix}
\usage{
as_r(x)
}
\arguments{
\item{x}{yacas list or list of lists to convert}
}
\description{
Convert yacas vector (list) or matrix (list of lists) to R vector/matrix
}
\concept{helper}
