% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_y}
\alias{as_y}
\title{Convert R vector/matrix to yacas vector (list) or matrix (list of lists)}
\usage{
as_y(x)
}
\arguments{
\item{x}{R vector to convert}
}
\description{
Convert R vector/matrix to yacas vector (list) or matrix (list of lists)
}
\concept{helper}
