% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{assume}
\alias{assume}
\title{Assume a variable has a value}
\usage{
assume(x, var, val)
}
\arguments{
\item{x}{yac_symbol}

\item{var}{Variable}

\item{val}{Value}
}
\description{
Assume a variable has a value
}
\concept{yac_symbol}
