% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{det}
\alias{det}
\title{Matrix Determinant}
\usage{
det(x, ...)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \code{\link[base:det]{base::det()}}.}

\item{\dots}{further arguments passed to \code{\link[base:det]{base::det()}}}
}
\description{
From \code{\link[base:det]{base::det()}}.
}
\examples{
(x <- matrix(1:4, ncol = 2))
det(x)
det(ysym(x))
}
\concept{yac_symbol}
