% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{diag<-}
\alias{diag<-}
\title{Matrix diagonals}
\usage{
diag(x) <- value
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \verb{base::diag<-()}.}

\item{value}{New value for \code{diag(x)}}
}
\description{
From \code{\link[base:diag]{base::diag()}}.
}
\concept{yac_symbol}
