% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{diag}
\alias{diag}
\title{Matrix diagonals}
\usage{
diag(x, ...)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \code{\link[base:diag]{base::diag()}}.}

\item{\dots}{further arguments passed to \code{\link[base:diag]{base::diag()}}}
}
\description{
From \code{\link[base:diag]{base::diag()}}.
}
\concept{yac_symbol}
