% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{equal_to}
\alias{equal_to}
\title{Add equality constraint for a yac symbol}
\usage{
equal_to(x, rhs)
}
\arguments{
\item{x}{A \code{yac_symbol}}

\item{rhs}{what \code{x} must be equal to}
}
\description{
Add equality constraint for a yac symbol
}
\concept{yac_symbol}
\concept{yac_symbol_solve}
