% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacas.R
\name{Eval}
\alias{Eval}
\alias{Eval.default}
\alias{Eval.Sym}
\alias{Eval.Expr}
\alias{Eval.yacas}
\title{Evaluate a yacas expression.}
\usage{
Eval(x, env = parent.frame(), ...)
}
\arguments{
\item{x}{Object to be evaluated.}

\item{env}{Environment or list in which to perform evaluation.}

\item{\dots}{Not currently used.}
}
\description{
Evaluate a yacas expression.
}
\examples{

Eval(yacas(expression(x*x)), list(x=2))

# same
x <- 2
Eval(yacas(expression(x*x)))

}
\keyword{symbolmath}
