% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacas.R
\name{get_xml_list_depth}
\alias{get_xml_list_depth}
\title{Matrix/vector:
Vector: List without nested lists
Matrix: List of Lists, but no futher nesting
Cannot be based on x, as this must work for e.g. Eval
as well. So need to be based on XML.
List in yacas is "<OMS cd="list1" name="list"/>"}
\usage{
get_xml_list_depth(x)
}
\description{
Matrix/vector:
Vector: List without nested lists
Matrix: List of Lists, but no futher nesting
Cannot be based on x, as this must work for e.g. Eval
as well. So need to be based on XML.
List in yacas is "<OMS cd="list1" name="list"/>"
}
