% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{\%*\%}
\alias{\%*\%}
\title{Matrix multiplication}
\usage{
x \%*\% y
}
\arguments{
\item{x}{A \code{yac_symbol}}

\item{y}{A \code{yac_symbol}}
}
\description{
Matrix multiplication
}
\concept{yac_symbol}
