% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{Hessian}
\alias{Hessian}
\title{Find the Hessian matrix of yac symbol}
\usage{
Hessian(expr, ...)
}
\arguments{
\item{expr}{A \code{yac_symbol}}

\item{\dots}{variables as character vector to take Hessian with respect to}
}
\description{
Find the Hessian matrix of yac symbol
}
\concept{yac_symbol}
