% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{integrate}
\alias{integrate}
\title{Integration of Functions}
\usage{
integrate(f, ...)
}
\arguments{
\item{f}{Function to integrate. See details.}

\item{\dots}{See details.}
}
\description{
If \code{f} is a \code{yac_symbol}, \code{yacas}'s \code{Integrate()} is used.
Else, \code{\link[stats:integrate]{stats::integrate()}} is used.
}
\details{
Additional arguments:
\itemize{
\item \code{\link{yac_symbol}}: \code{var}, \code{lower}, \code{upper}
\item Else (\code{\link[stats:integrate]{stats::integrate()}}): \code{lower}, \code{upper}
}
}
\concept{yac_symbol}
