% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{Jacobian}
\alias{Jacobian}
\title{Find the Jacobian matrix of yac symbol}
\usage{
Jacobian(expr, vars)
}
\arguments{
\item{expr}{A \code{yac_symbol}}

\item{name}{variables to take Hessian with respect to}
}
\description{
Find the Jacobian matrix of yac symbol
}
\concept{yac_symbol}
