% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{lim}
\alias{lim}
\title{Limits}
\usage{
lim(...)
}
\arguments{
\item{\dots}{See details.}
}
\description{
If first argument is a \code{yac_symbol}, \code{yacas}'s \code{Limit()} is used.
}
\details{
Arguments:
\itemize{
\item \code{\link{yac_symbol}}: \code{f}, \code{var}, \code{val}, \code{from_left}, \code{from_right}
}
}
\concept{yac_symbol}
