% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{Math.yac_symbol}
\alias{Math.yac_symbol}
\title{Math functions}
\usage{
\method{Math}{yac_symbol}(x, ...)
}
\arguments{
\item{x}{\code{yac_symbol}.}

\item{\dots}{further arguments passed to methods}
}
\description{
Math functions
}
