% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{matrix_product}
\alias{matrix_product}
\title{Matrix product of two matrices}
\usage{
matrix_product(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}
}
\description{
Matrix product of two matrices
}
