% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{Ops.yac_symbol}
\alias{Ops.yac_symbol}
\title{Math operators}
\usage{
\method{Ops}{yac_symbol}(e1, e2)
}
\arguments{
\item{e1}{A \code{yac_symbol}.}

\item{e2}{A \code{yac_symbol}.}
}
\description{
Math operators
}
