% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{pow}
\alias{pow}
\alias{pow.default}
\title{Matrix Power}
\usage{
pow(x, n, ...)

\method{pow}{default}(x, n, ...)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \code{\link[=pow.default]{pow.default()}}.}

\item{n}{\code{n}th power of the square matrix.}

\item{\dots}{further arguments passed to \code{\link[=pow.default]{pow.default()}}}
}
\description{
Matrix Power
}
\examples{
(x <- matrix(c(1, 2, 2, 3), ncol = 2))
pow(x, 2)
pow(ysym(x), 2)
}
\concept{yac_symbol}
