% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{prod.yac_symbol}
\alias{prod.yac_symbol}
\title{Product of Vector Elements}
\usage{
\method{prod}{yac_symbol}(expr, ..., na.rm = FALSE)
}
\arguments{
\item{expr}{Expression to be multiplied}

\item{\dots}{Not used}

\item{na.rm}{Not used}
}
\description{
Product of Vector Elements
}
\concept{yac_symbol}
