% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{rbind.yac_symbol}
\alias{rbind.yac_symbol}
\title{Combine R Objects by Rows}
\usage{
\method{rbind}{yac_symbol}(..., deparse.level = 1)
}
\arguments{
\item{\dots}{Objects to bind}

\item{deparse.level}{Not used}
}
\description{
Combine R Objects by Rows
}
\concept{yac_symbol}
