% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryacas-package.R
\docType{package}
\name{Ryacas-package}
\alias{Ryacas}
\alias{Ryacas-package}
\title{R interface to yacas computer algebra package}
\description{
Ryacas allows one to use the yacas computer algebra package entirely from
within R.
}
\details{
Please read the "Getting started" vignette.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-cas/ryacas}
  \item \url{http://www.yacas.org}
  \item Report bugs at \url{https://github.com/r-cas/ryacas/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Meyer Andersen \email{mikl@math.aau.dk} [copyright holder]

Authors:
\itemize{
  \item Rob Goedman \email{goedman@mac.com} [copyright holder]
  \item Gabor Grothendieck \email{ggrothendieck@gmail.com} [copyright holder]
  \item Søren Højsgaard \email{sorenh@math.aau.dk} [copyright holder]
  \item Grzegorz Mazur \email{teoretyk@gmail.com} [copyright holder]
  \item Ayal Pinkus \email{apinkus@xs4all.nl} [copyright holder]
}

Other contributors:
\itemize{
  \item Nemanja Trifunovic (UTF-8 part of yacas (src/yacas/include/yacas/utf8*)) [copyright holder]
}

}
\keyword{programming}
