% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{Ryacas_options}
\alias{Ryacas_options}
\title{Set or get options for the Ryacas package}
\usage{
Ryacas_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key] = [value]} pairs to set options.}
}
\description{
Set or get options for the Ryacas package
}
\section{Supported options}{

The following options are supported
\itemize{
\item \code{module_matvec_enabled} (default \code{TRUE}): Print yacas \code{List()}'s as
vectors and \code{List(List(), ...)}'s as matrices.
\item \code{prettyform_default} (default \code{FALSE}): Print yacas as \code{PrettyForm()} as default.
}
}

