% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_output_width}
\alias{set_output_width}
\title{Set width of yacas output}
\usage{
set_output_width(w)
}
\arguments{
\item{w}{Width in number of characters}
}
\description{
Set width of yacas output
}
\concept{yac_communication}
\keyword{internal}
