% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{simplify}
\alias{simplify}
\title{Simplify expression}
\usage{
simplify(x, timeout = 2)
}
\arguments{
\item{x}{A \code{yac_symbol}}

\item{timeout}{timeout in seconds before simplification is aborted;
only works when package unix is available}
}
\description{
Simplify expression
}
\concept{yac_symbol}
