% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{solve.yac_symbol}
\alias{solve.yac_symbol}
\title{Solve a system of equations}
\usage{
\method{solve}{yac_symbol}(a, b, ...)
}
\arguments{
\item{a}{If \code{yac_symbol} treat as such, else
call \code{\link[base:solve]{base::solve()}}.}

\item{b}{If \code{yac_symbol} treat as such, else
call \code{\link[base:solve]{base::solve()}}.}

\item{\dots}{Not used}
}
\description{
This generic function solves the
equation $a %*% x = b$ for $x$.
}
