% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expr.R
\name{stripvar}
\alias{stripvar}
\title{Removes part of expression containing variable}
\usage{
stripvar(expr, var)
}
\arguments{
\item{expr}{Expression where \code{x == expr} should be replaced to \code{expr}}

\item{var}{Name of variable, e.g. \code{x}}
}
\description{
Yacas' \code{Solve(eq, x)} can return e.g.
\code{x == expr} and \code{{x == expr1, x == expr2, ...}}.
Some usages are easier if the initial \code{x == } part is removed.
This is the purpose of this function.
}
\keyword{symbolmath}
