% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{[[.yac_symbol}
\alias{[[.yac_symbol}
\title{Extract parts of an object}
\usage{
\method{[[}{yac_symbol}(x, i)
}
\arguments{
\item{x}{A \code{yac_symbol}.}

\item{i}{indices specifying elements to extract}
}
\description{
Extract parts of an object
}
