% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{[<-.yac_symbol}
\alias{[<-.yac_symbol}
\title{Extract or replace parts of an object}
\usage{
\method{[}{yac_symbol}(x, i, j) <- value
}
\arguments{
\item{x}{A \code{yac_symbol}.}

\item{i}{row indices specifying elements to extract or replace}

\item{j}{column indices specifying elements to extract or replace}

\item{value}{the value to replace \code{x[i, j]} by}
}
\description{
Extract or replace parts of an object
}
