% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{sum}
\alias{sum}
\title{Summation of Functions}
\usage{
sum(...)
}
\description{
If first argument is a \code{yac_symbol}, \code{yacas}'s \code{Sum()} is used.
Else, \code{\link[base:sum]{base::sum()}} is used.
}
\details{
Additional arguments:
\itemize{
\item \code{\link{yac_symbol}}: \code{f}, \code{var}, \code{lower}, \code{upper}
}
}
\concept{yac_symbol}
