% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{sum.yac_symbol}
\alias{sum.yac_symbol}
\title{Summation}
\usage{
\method{sum}{yac_symbol}(expr, var, lower, upper, ..., na.rm = FALSE)
}
\arguments{
\item{expr}{Expression to be summed}

\item{var}{Variable to sum}

\item{lower}{Lower limit}

\item{upper}{Upper limit}

\item{\dots}{Not used}

\item{na.rm}{Not used}
}
\description{
If only \code{expr} given: sum elements.
}
\details{
Else: sums \code{expr} by letting \code{var} taking values
from \code{lower} to \code{upper} (potentially \code{Inf})
}
\concept{yac_symbol}
