% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacas.R
\name{syacas}
\alias{syacas}
\title{yacas interface -- silent version}
\usage{
syacas(x, ...)
}
\arguments{
\item{x}{A yacas character string or an R expression without terminating
semicolon to be processed by yacas.}

\item{...}{Additional arguments ultimately passed down to
\code{yacas.character}.}
}
\description{
Similar to \code{\link[=yacas]{yacas()}} but silent.
This can be useful when working with yacas directly.
}
\seealso{
yacas
}
