\name{Sym}
\alias{Sym}

%- Also NEED an '\alias' for EACH other topic documented here.

\alias{as.character.Sym}
\alias{Deriv}
\alias{Integrate}
\alias{OpenMath2R}
\alias{Ops.Sym}
\alias{Ops.yacas.symbol}
\alias{print.Sym}
\alias{print.yacas}
\alias{Sym}
\alias{SymExpr}
\alias{trans}
\alias{transtab}
\alias{yacas.symbol.value}
\alias{yDeriv}
\alias{yFactorial}
\alias{yIntegrate}
\alias{yLimit}
\alias{yrewrite}
\alias{yUnlist}

\alias{Sin}
\alias{Cos}
\alias{Tan}
\alias{ArcSin}
\alias{ArcCos}
\alias{ArcTan}
\alias{Log}
\alias{Exp}
\alias{Simplify}
\alias{Factorial}
\alias{List}
\alias{Ver}
\alias{Seq}
\alias{N}
\alias{Pi}
\alias{Clear}
\alias{Factor}
\alias{Expand}
\alias{Taylor}
\alias{InverseTaylor}
\alias{PrettyForm}
\alias{TeXForm}
\alias{Precision}
\alias{Conjugate}
\alias{PrettyPrinter}
\alias{Solve}
\alias{Newton}
\alias{Set}
\alias{Infinity}
\alias{I}
\alias{x0}
\alias{x1}
\alias{x2}
\alias{x3}
\alias{x4}
\alias{x5}
\alias{x6}
\alias{x7}
\alias{x8}
\alias{x9}
\alias{x10}
\alias{Limit}
\alias{Subst}
\alias{Inverse}
\alias{Determinant}

\title{ Sym }
\description{
  The Symbol interface to yacas.
}
\usage{
Sym(...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{ An R character string or object that can be coerced to
a character string.}
}
\details{
  See the vignette, Ryacas-Sym, for numerous examples.
}
\value{
Returns a yacas \code{Sym} object.
}
\examples{
}
\keyword{ symbolmath }% at least one, from doc/KEYWORDS
