% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{t}
\alias{t}
\alias{t.yac_symbol}
\title{t}
\usage{
\method{t}{yac_symbol}(x)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else
call \code{\link[base:t]{base::t()}}.}
}
\description{
t
}
\concept{yac_symbol}
