% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{tr}
\alias{tr}
\alias{tr.default}
\title{Matrix Trace}
\usage{
tr(x, ...)

\method{tr}{default}(x, ...)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \code{\link[=tr.default]{tr.default()}}.}

\item{\dots}{further arguments passed to \code{\link[=tr.default]{tr.default()}}}
}
\description{
The trace of a square matrix is the sum of the diagonal elements.
}
\examples{
(x <- matrix(1:4, ncol = 2))
tr(x)
tr(ysym(x))
}
\concept{yac_symbol}
