% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{upper.tri}
\alias{upper.tri}
\title{Lower and upper triangular part of a matrix}
\usage{
upper.tri(x, diag = FALSE)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else
call \code{\link[base:lower.tri]{base::lower.tri()}}/\code{\link[base:lower.tri]{base::upper.tri()}}.}

\item{diag}{Whether diagonal is included.}
}
\description{
Lower and upper triangular part of a matrix
}
\concept{yac_symbol}
