% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{vec}
\alias{vec}
\alias{vec.default}
\title{Vectorize}
\usage{
vec(x, ...)

\method{vec}{default}(x, ...)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \code{\link[base:vector]{base::as.vector()}}.}

\item{\dots}{further arguments passed to \code{\link[base:vector]{base::as.vector()}}}
}
\description{
Vectorize
}
\examples{
(x <- matrix(1:9, ncol = 3))
vec(x)
vec(ysym(x))
}
\concept{yac_symbol}
