% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{vech}
\alias{vech}
\alias{vech.default}
\title{Half-Vectorize}
\usage{
vech(x, ...)

\method{vech}{default}(x, ...)
}
\arguments{
\item{x}{If \code{yac_symbol} treat as such, else call \code{\link[=vech.default]{vech.default()}}.}

\item{\dots}{further arguments passed to \code{\link[=vech.default]{vech.default()}}}
}
\description{
Half-Vectorize
}
\examples{
A <- mtcars[, c(1, 3, 4, 5, 6, 7)]
x <- cov(A)
vech(x)
vech(ysym(x))
}
\concept{yac_symbol}
