% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{with_value}
\alias{with_value}
\title{Give a variable a value}
\usage{
with_value(x, var, val)
}
\arguments{
\item{x}{yac_symbol}

\item{var}{Variable}

\item{val}{Value}
}
\description{
Give a variable a value
}
\concept{yac_symbol}
