% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-fn.R
\name{y_fn}
\alias{y_fn}
\title{Prepare simple \code{yacas} call}
\usage{
y_fn(x, fn, ...)
}
\arguments{
\item{x}{parameter to function \code{fn}}

\item{fn}{function with parameter \code{x}}

\item{\dots}{additional arguments to \code{fn}}
}
\description{
Prepare simple \code{yacas} call
}
\examples{
y_fn("x^2 - 1", "TeXForm")
yac_str(y_fn("x^2 - 1", "TeXForm"))

y_fn("x^2 - 1", "Factor")
yac_str(y_fn("x^2 - 1", "Factor"))

cmd <- "x^2 - 1 == 0" \%>\% y_fn("Solve", "x")
cmd
sol <- yac_str(cmd)
sol
yac_str(y_rmvars(sol))

}
\concept{helper}
\concept{yac_symbol}
