% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{y_print}
\alias{y_print}
\title{Pretty print yacas strings}
\usage{
y_print(x)
}
\arguments{
\item{x}{yacas string, e.g. a matrix}
}
\description{
Pretty print yacas strings
}
\examples{
A <- diag(4)
Ayac <- as_y(A)
y_print(Ayac)

B <- A
B[2, 2] <- "-t"
Byac <- as_y(B)
Byac
y_print(Byac)

}
\concept{helper}
