% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-funcs.R
\name{yac}
\alias{yac}
\title{Run yacas command}
\usage{
yac(x, return = "str")
}
\arguments{
\item{x}{yacas command}

\item{rettype}{\code{str} for string/character, \code{expr} for expression, \code{silent} for silent}
}
\description{
Run yacas command
}
\concept{yac_communication}
