% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-funcs.R
\name{yac_assign}
\alias{yac_assign}
\title{Assign yacas variable}
\usage{
yac_assign(value, x)
}
\arguments{
\item{value}{Expression}

\item{x}{Variable name}
}
\description{
Assign yacas variable
}
\concept{yac_communication}
\concept{yac_symbol}
