% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-cli.R
\name{yac_cli}
\alias{yac_cli}
\title{yacas command line interface}
\usage{
yac_cli(enable_history = TRUE)
}
\arguments{
\item{enable_history}{Use \code{R} history such that previous yacas commands can be used. Default is \code{TRUE}.}
}
\value{
Output of yacas is returned.
}
\description{
Interactive interface to the yacas
}
\details{
The user types valid yacas input and presses return. Type 'quit' to return
to R prompt.
}
\note{
Note that command will use R \code{history()} and modify it by default.
Yacas is given a limited amount of time to complete, otherwise
\verb{\[1\] CommandLine(1) : User interrupted calculation} is returned. E.g.
Taylor(x,0,5) 1/(1+x) will work, but Taylor(x,0,12) 1/(1+x) is likely to
take too long.
}
\examples{

\dontrun{
yac_cli()
 (x+y)^3-(x-y)^3
 Simplify(\%)
 q
}

}
\references{
\url{https://yacas.sourceforge.io/}
}
\concept{yac_communication}
\keyword{symbolmath}
