% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{yac_core}
\alias{yac_core}
\title{Evaluate yacas expression}
\usage{
yac_core(expr)
}
\arguments{
\item{expr}{Yacas expression}
}
\value{
Result of evaluating \code{expr} by yacas and
side-effects of the evaluation
}
\description{
This is a low-level function for evaluating yacas expression represented as
string.
}
\examples{
yac_core("D(x)Sin(x^2)")

}
\concept{yac_communication}
\keyword{internal}
