% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-funcs.R
\name{yac_expr}
\alias{yac_expr}
\title{Run yacas command returning R expression}
\usage{
yac_expr(x)
}
\arguments{
\item{x}{yacas command}
}
\description{
Run yacas command returning R expression
}
\concept{yac_communication}
