% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-funcs.R
\name{yac_silent}
\alias{yac_silent}
\title{Run yacas command silently}
\usage{
yac_silent(x)
}
\arguments{
\item{x}{yacas command}
}
\description{
Run yacas command silently
}
\concept{yac_communication}
\concept{yac_symbol}
