% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{yac_solve_str}
\alias{yac_solve_str}
\title{Run yacas Solve() command returning string/character}
\usage{
yac_solve_str(x)
}
\arguments{
\item{x}{yacas command starting with a Solve()}
}
\description{
Run yacas Solve() command returning string/character
}
\concept{solve}
