% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-funcs.R
\name{yac_str}
\alias{yac_str}
\title{Run yacas command returning string/character}
\usage{
yac_str(x)
}
\arguments{
\item{x}{yacas command}
}
\description{
Run yacas command returning string/character
}
\concept{yac_communication}
