% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{yac_symbol}
\alias{yac_symbol}
\title{Make a yacas symbol}
\usage{
yac_symbol(x)
}
\arguments{
\item{x}{A vector or a matrix}
}
\value{
A \code{yac_symbol}
}
\description{
This is an alias for \code{\link[=ysym]{ysym()}}.
See description there.
}
\concept{yac_symbol}
