% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{ysym_make}
\alias{ysym_make}
\title{Construct and assign a vector of variables}
\usage{
ysym_make(x, overwrite = FALSE, warn = TRUE)
}
\arguments{
\item{x}{Vector of variable names. If named, the names are the variables
and the content is the value.}

\item{overwrite}{overwrite if exists already}

\item{warn}{warn if name already exists in \code{.GlobalEnv}}
}
\value{
Invisibly names assigned
}
\description{
If unnamed vector, the content is the variables.
}
\details{
If named vector, the names are the variables,
and the values are the content.

Note, that it has the side effect that it assigns variables in
\code{.GlobalEnv}.
}
\examples{
ysym_ls()
ysym_make(c("a", "b"))
ysym_ls()
a

ysym_make(c("xt" = "{{1, t}}", "xtk" = "{{1, t+k}}"))
ysym_ls()
xt
ysym_ls(print_details = TRUE)

# Warning
# ysym_make(c("a", "b"))
ysym_make(c("a" = "2", "b" = "3"), overwrite = TRUE, warn = FALSE)
a

}
