% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sym2.R
\name{as.Sym.character}
\alias{as.Sym.character}
\title{Convert character vector to yacas object}
\usage{
\method{as.Sym}{character}(x, ...)
}
\arguments{
\item{x}{An R character vector.}

\item{\dots}{Not used}
}
\description{
Simple and raw conversion to yacas
}
\examples{
x <- c("a", "2", "4", "c", "d", "6")
x
y <- as.Sym(x)
y
Eval(y, list(a = 3, c = 3, d = 3))

}
