% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sym2.R
\name{as.Sym.matrix}
\alias{as.Sym.matrix}
\title{Convert character matrix to yacas object}
\usage{
\method{as.Sym}{matrix}(x, ...)
}
\arguments{
\item{x}{An R character matrix.}

\item{\dots}{Not used}
}
\description{
Simple and raw conversion to yacas
}
\examples{
x <- matrix(c("a", "2", "4", "c", "d", "6"), 3, 2)
x
y <- as.Sym(x)
y
Eval(y, list(a = 3, c = 3, d = 3))

}
