% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{as_r}
\alias{as_r}
\title{Convert yacas object to R}
\usage{
as_r(x)
}
\arguments{
\item{x}{yacas list or list of lists to convert}
}
\description{
If \code{x} is a yacas command as string, convert to a character vector/matrix in R.
If \code{x} is a \code{yac_symbol} (e.g. from \code{\link[=yac_symbol]{yac_symbol()}}), then convert it to a numeric object
if there are no variables or a character type if there are variables.
}
\details{
In yacas a vector is a list, and a matrix is a list of lists.
}
\concept{helper}
