% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{deriv.yac_symbol}
\alias{deriv.yac_symbol}
\title{Find the derivative of yac symbol}
\usage{
\method{deriv}{yac_symbol}(expr, ...)
}
\arguments{
\item{expr}{A \code{yac_symbol}}

\item{\dots}{variables as character vector to take derivate with respect to}
}
\description{
Find the derivative of yac symbol
}
\concept{yac_symbol}
