% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryacas-package.R
\docType{package}
\name{Ryacas-package}
\alias{Ryacas-package}
\title{R interface to yacas computer algebra package}
\description{
Ryacas allows one to use the yacas computer algebra package entirely from
within R.
}
\details{
Please read the "Getting started" vignette.
}
\keyword{programming}
