% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-fn.R
\name{y_eval}
\alias{y_eval}
\title{Evaluate a yacas expression}
\usage{
y_eval(expr, varlist)
}
\arguments{
\item{expr}{a valid yacas expression}
}
\description{
Evaluate a yacas expression by replacing variables with values as for the
given list.
}
\examples{
eq <- ysym("2*y+x^2+2*x-3")
y_eval(eq, list(x=3, y=2))

}
\concept{helper}
