% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y-fn.R
\name{y_rmvars}
\alias{y_rmvars}
\title{Remove/strip variable names}
\usage{
y_rmvars(x)
}
\arguments{
\item{x}{yacas command}
}
\description{
This only builds a yacas command.
You need to also call \code{\link[=yac_str]{yac_str()}}, \code{\link[=yac_expr]{yac_expr()}} or similar.
This is the reason that it does not call yacas: it depends on how you
want it returned (string, expression).
}
\examples{
cmd <- "{x == 2, y == 4}"
yac_str(cmd)
yac_str(y_rmvars(cmd))

}
\concept{helper}
