% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-funcs.R
\name{yac}
\alias{yac}
\title{Run yacas command}
\usage{
yac(x, rettype = c("str", "expr", "silent"))
}
\arguments{
\item{x}{yacas command}

\item{rettype}{\code{str} for string/character, \code{expr} for expression, \code{silent} for silent}
}
\description{
Run yacas command
}
\examples{
yac("D(x) x^2 + 4*x")
yac("D(x) x^2 + 4*x", rettype = "str")
yac("D(x) x^2 + 4*x", rettype = "expr")
yac("D(x) x^2 + 4*x", rettype = "silent")

}
\concept{yac_communication}
\concept{yac_symbol}
