% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yac-symbol.R
\name{yac_symbol}
\alias{yac_symbol}
\title{Make a yacas symbol}
\usage{
yac_symbol(x)
}
\arguments{
\item{x}{A vector or a matrix}
}
\value{
A \code{yac_symbol}
}
\description{
Note that this results in multiple calls to \code{yacas} and
the performance may be slower than manually using e.g. \code{\link[=yac_str]{yac_str()}}.
}
\concept{yac_symbol}
