% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yacmode.R
\name{yacmode}
\alias{yacmode}
\title{yacmode interface}
\usage{
yacmode()
}
\value{
Output of yacas is returned.
}
\description{
Interactive interface to the yacas
}
\details{
The user types valid yacas input and presses return. Type 'quit' to return
to R prompt.
}
\note{
Note that command recall will recall previous R commands, not previous
yacas input. Yacas is given a limited amount of time to complete, otherwise
\code{\[1\] CommandLine(1) : User interrupted calculation} is returned. E.g.
Taylor(x,0,5) 1/(1+x) will work, but Taylor(x,0,12) 1/(1+x) is likely to
take too long.
}
\examples{

\dontrun{
yacmode()
 (x+y)^3-(x-y)^3
 Simplify(\%)
 q
}

}
\references{
\url{http://yacas.sourceforge.net}
}
\keyword{symbolmath}
